chore(test): cleanup
